+++ src/iterators/version.rs
@@ -373,7 +373,7 @@ pub(crate) mod raw {
         unsafe fn translated_desc(self: &VerIterator) -> UniquePtr<DescIterator>;
 
         #[cxx_name = "Index"]
-        pub fn index(self: &VerIterator) -> u64;
+        pub fn index(self: &VerIterator) -> usize;
         /// Clone the pointer.
         ///
         /// # Safety
